Moved all simulation data loading classes to data package.
